% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_primary_color_is_ok.R
\name{check_primary_color_is_ok}
\alias{check_primary_color_is_ok}
\title{Check if your selection it is in squirrels' fur color vector}
\usage{
check_primary_color_is_ok(string)
}
\arguments{
\item{string}{You need to input as the parameter a string color}
}
\value{
A boolean: TRUE if your string it is in fur color vector. FALSE otherwise
}
\description{
This function aims to evaluate whether a string character corresponds or not to the squirrels' fur color vector
}
\examples{
check_primary_color_is_ok("White")
}
