% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_message_fur_color.R
\name{get_message_fur_color}
\alias{get_message_fur_color}
\title{Get a message with a fur color}
\usage{
get_message_fur_color(primary_fur_color)
}
\arguments{
\item{primary_fur_color}{Character. The primary fur color of interest}
}
\value{
None. Used for side effect. Outputs a message in the console
}
\description{
You can get a message with the fur color of interest with \code{get_message_fur_color()}
}
\examples{
get_message_fur_color(primary_fur_color = "Cinnamon")
}
